+++ prc_PLC.py
@@ -35,6 +35,8 @@ def start_process(config_file):
     __ini_speed_H = __ini_prc_config__.plc_info.ini_speed_H
     __ini_speed_L = __ini_prc_config__.plc_info.ini_speed_L
     __ini_conv_length = __ini_prc_config__.plc_info.ini_conv_length
+    
+    __ini_start_conv = __ini_prc_config__.Config.StartConv
 
     # 连接PLC
     inst_logger.info(f"线程{__prc_name__}尝试连接 PLC")
@@ -90,6 +92,8 @@ def start_process(config_file):
     
     lst_parcel_keys =[]
     dict_reading_con = {}
+    if __ini_start_conv:
+        inst_redis.setkey("plc_conv:command","start")
     while b_thread_running:
         # 刷新当前线程的运行锁
         inst_redis.setkeypx(f"pro_mon:{__prc_name__}:run_lock", __prc_id__, __prc_expiretime)
@@ -140,12 +144,14 @@ def start_process(config_file):
 
         # 减速到高速
         if plc_conv_status == 'run' and plc_conv_fullspeed == 'Yes' and is_locked is True:
+            is_locked = False
             for i in range(1, 5):
                 __inst_plc__.db_write(12, 3 + (i - 1) * 4, bytearray([3]))
             for i in range(1, 5):
                 inst_redis.setkey(f"plc_conv:CV0{i}:speed", "high")
             # inst_redis.lpush("stm_sys_log", f"[{current_time}]-<conv>: slow down ——》 high")
             inst_logger.info("线程 %s 皮带机从减速变为高速，速度为高速" % (__prc_name__,))
+            inst_redis.setkey(f"plc_conv:fullspeed", "Yes")
         # 减速变成停止
         if plc_conv_status == 'run' and plc_conv_fullspeed is None:
             is_locked = False
@@ -198,12 +204,13 @@ def start_process(config_file):
                         
                         # 清理redis中CV03上的包裹信息
                         inst_redis.clearkey(f"{key}")
-                        inst_redis.clearkey(f"parcel:posy:{lst_splited_keys[-1]}")
-                        inst_redis.clearkey(f"parcel:sid:{lst_splited_keys[-1]}")
-                        inst_redis.clearkey(f"parcel:scan_result:{lst_splited_keys[-1]}")
-                        inst_redis.clearkey(f"parcel:barcode:{lst_splited_keys[-1]}")
+                        inst_redis.clearkey(f"parcel:posy:{str_uid}")
+                        inst_redis.clearkey(f"parcel:sid:{str_uid}")
+                        inst_redis.clearkey(f"parcel:scan_result:{str_uid}")
+                        inst_redis.clearkey(f"parcel:barcode:{str_uid}")
                         
                         # 将 str_barcode 从set_reading_gr中删除
+                        # 此处需要深入考虑一下，如果单次扫描可以使用gr来剔除重复扫描，那么可以考虑暂时不删除，等到切换MAWB在清理
                         # inst_redis 
                         inst_logger.debug(f"包裹已经离开CV03,uid={str_uid},barcode={str_barcode}") 
             
@@ -232,6 +239,8 @@ def start_process(config_file):
         prc_run_lock = inst_redis.getkey(f"pro_mon:{__prc_name__}:run_lock")
         if prc_run_lock is None:
             int_exit_code = 1
+            for i in range(1, 5):
+                __inst_plc__.db_write(12, 3 + (i - 1) * 4, bytearray([0]))
             break
 
         # 如command区收到退出命令，根据线程类型决定是否立即退出
@@ -240,6 +249,8 @@ def start_process(config_file):
             # 在此处判断是否有尚未完成的任务，或尚未处理的stm序列；
             # 如有则暂缓退出，如没有立即退出
             int_exit_code = 2
+            for i in range(1, 5):
+                __inst_plc__.db_write(12, 3 + (i - 1) * 4, bytearray([0]))
             break
 
     inst_logger.info("线程 %s 已退出，返回代码为 %d" % (__prc_name__, int_exit_code))